pyproject.toml
[project]
name = "echofinder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "charset-normalizer>=3.4.7",
    "pillow>=12.2.0",
    "platformdirs>=4.9.6",
    "pygments>=2.20.0",
    "pymupdf>=1.27.2.2",
    "pyqt6>=6.11.0",
    "python-magic>=0.4.27",
    "python-vlc>=3.0.21203",
]
